refactor(ui): remove the Pixi renderer — commit fully to Three.js
The Three.js renderer is now the sole graph canvas, so drop the legacy Pixi
renderer and its dual-renderer plumbing.

- Delete PixiGraphCanvas, PixiRenderer, usePixiLayout, spriteTextures,
  PixiControlPanel, pixiLayoutWorker, and the Pixi-only playground stories.
- Move the renderer-agnostic helpers (scaleBreakpoints, viewport) out of
  pixi/ into three/, where the only remaining consumers live.
- GraphViewer always uses ThreeGraphCanvas (drop the ?renderer=pixi switch).
- Barrel: Graph / GraphCanvas now resolve to ThreeGraphCanvas; LayoutMode and
  the breakpoint exports re-point to the Three sources.
- Drop the pixi.js dependency.
